Refactor: Mover data class `Disco` a su propio archivo y actualizar referencias
Se ha movido la data class `Disco` del archivo `DiscosRecyclerView.kt` a un nuevo archivo `model/Disco.kt`.

Además, se han realizado los siguientes cambios:

- **Renombrado de archivo:** `DiscosRecyclerView.kt` ha sido renombrado a `DiscoAdapter.kt` para reflejar mejor su contenido.
- **Actualización de importaciones:** Se han actualizado las importaciones en `DiscosActivity.kt` y `DiscoAdapter.kt` para apuntar a la nueva ubicación de la clase `Disco`.
- **Comentarios y organización:** Se han añadido y mejorado comentarios en `DiscoAdapter.kt` y `DiscosActivity.kt` para una mejor comprensión del código. En `DiscosActivity.kt`, se ha reorganizado el código separando las funciones de inicialización de componentes.
